Split sequence building vs sending
