+++ config/config.debug.toml
@@ -31,7 +31,8 @@ SyncChunkSize = 100
 TrustedSequencerURI = ""
 
 [Sequencer]
-WaitPeriodPoolIsEmpty = "15s"
+WaitPeriodPoolIsEmpty = "1s"
+WaitPeriodSendSequence = "15s"
 LastBatchVirtualizationTimeMaxWaitPeriod = "300s"
 WaitBlocksToUpdateGER = 10
 LastTimeBatchMaxWaitPeriod = "15s"
