+++ sequencer/config.go
@@ -7,8 +7,11 @@ import (
 
 // Config represents the configuration of a sequencer
 type Config struct {
+	// WaitPeriodSendSequence is the time the sequencer waits until
+	// trying to send a sequence to L1
+	WaitPeriodSendSequence types.Duration `mapstructure:"WaitPeriodSendSequence"`
 	// WaitPeriodPoolIsEmpty is the time the sequencer waits until
-	// trying to propose a batch
+	// trying to add new txs to the state
 	WaitPeriodPoolIsEmpty types.Duration `mapstructure:"WaitPeriodPoolIsEmpty"`
 
 	// LastBatchVirtualizationTimeMaxWaitPeriod is time since sequences should be sent
